refactor index fixture into module package
