+++ test/fixture/ast/src/commonMain/kotlin/io/github/charlietap/chasm/fixture/type/HeapType.kt
@@ -3,7 +3,7 @@ package io.github.charlietap.chasm.fixture.type
 import io.github.charlietap.chasm.ast.module.Index
 import io.github.charlietap.chasm.ast.type.AbstractHeapType
 import io.github.charlietap.chasm.ast.type.ConcreteHeapType
-import io.github.charlietap.chasm.fixture.instruction.typeIndex
+import io.github.charlietap.chasm.fixture.module.typeIndex
 
 fun heapType() = abstractHeapType()
